Pull from Dev to Players (#48)
* Update act.movement.c (#35)

* Update act.movement.c

For some reason this buffer was allocated with size 512, where in another location a description was fread() into a buffer of size 1024 (SMALL_BUFSIZE) which pointer was assigned to short_description. No reason to be stingy.

* Elimate the need for a buffer

---------

Co-authored-by: Doug Johnson <doug@eq0.co>

* Remove debuglog (#32)

* add gold reimb for TR (#41)

* Add basic mssp support (#30)

* Fix for do_put strcpy() overwrite (#36)

* Fix for do_put strcpy() overwrite

Changed strcpy() to memmove() to fix overlapping memory.

* Calculate string length correctly

---------

Co-authored-by: Doug Johnson <doug@eq0.co>

* Remove separate listing of commands for builders port (#43)

* Update structs.h

Add master flag for mobs (for some reason the last bit of mob flags was not set. Rectified!)

* Clan improvements (#45)

1. Add clan donation room to the clans file,
2. Add clan charter/description to the clans file.,
3. Add functions to add/change both of those. For imms and clanleaders for the charters, and just imms for the clan donation room.,
4. Allow all players to see clan info with claninfo command: they can see the name of the clan, it's charter, leader, and members. Imms can see more info, like vnums of clan recall room, clan piece, and donation room.,
5. Add a few items to listclans for imms, moved it around so clan number is all the way on the left.

* Update clan.c

Fixed reference to null pointer.

* Update clan.c

Fixed do_charter to match imm level restrictions.

* Update structs.h

Added mob flags:
 MOB_WEAPONSMITH - works with weapons
 MOB_ARMORER - works with armor
 MOB_JEWELER - works with accessories/jewelry

These are in addition to but not solely for the MOB_MASTER flag.

* Update constants.c

Added bit descriptions for:
Player is editing the charter (don't set), and
Mobs that is a master, weaponsmith, armorer, and jeweler.

* Update olc.h

Increased sizes of MOB and MOB2 flags so they show in medit.

* Master repairman (#47)

* Update utils.c

Price adjust fix to remove character alignment from calculation.
Also added a description to the function.

* Update act.comm.c

Fixed an issue where a mobs short description would not get capitalized when telling you something.

* Update spec_procs.c

Added master and specialized smiths to repair_guy code.
Added messages that match the repaired item quality.

---------

Co-authored-by: Doug Johnson <doug@eq0.co>